+++ amethyst_assets/src/prefab/mod.rs
@@ -34,11 +34,15 @@ pub trait PrefabData<'a> {
     /// - `system_data`: `SystemData` needed to do the loading
     /// - `entities`: Some components need access to the entities that was created as part of the
     ///               full prefab, for linking purposes, so this contains all those `Entity`s.
+    /// - `children`: Entities that need access to the `Hierarchy`  in this function won't be able
+    ///               to access it yet, since it is only updated after this function runs. As a work-
+    ///               around, this slice includes all hierarchical children of the entity being passed.
     fn add_to_entity(
         &self,
         entity: Entity,
         system_data: &mut Self::SystemData,
         entities: &[Entity],
+        children: &[Entity],
     ) -> Result<Self::Result, Error>;
 
     /// Trigger asset loading for any sub assets.
@@ -358,6 +362,7 @@ where
         &self,
         entity: Entity,
         system_data: &mut Self::SystemData,
+        _: &[Entity],
         _: &[Entity],
     ) -> Result<Handle<A>, Error> {
         let handle = match *self {