+++ amethyst_derive/src/prefab_data.rs
@@ -24,6 +24,7 @@ fn impl_prefab_data_component(ast: &DeriveInput) -> TokenStream {
             fn add_to_entity(&self,
                              entity: Entity,
                              system_data: &mut Self::SystemData,
+                             _: &[Entity],
                              _: &[Entity]) -> ::std::result::Result<(), Error> {
                 system_data.insert(entity, self.clone()).map(|_| ())?;
                 Ok(())
@@ -56,7 +57,7 @@ fn impl_prefab_data_aggregate(ast: &DeriveInput) -> TokenStream {
             }
         } else {
             quote! {
-                self.#name.add_to_entity(entity, &mut system_data.#tuple_index, entities)?;
+                self.#name.add_to_entity(entity, &mut system_data.#tuple_index, entities, children)?;
             }
         }
     });
@@ -88,7 +89,8 @@ fn impl_prefab_data_aggregate(ast: &DeriveInput) -> TokenStream {
             fn add_to_entity(&self,
                              entity: Entity,
                              system_data: &mut Self::SystemData,
-                             entities: &[Entity]) -> ::std::result::Result<(), Error> {
+                             entities: &[Entity],
+                             children: &[Entity]) -> ::std::result::Result<(), Error> {
                 #(#adds)*
                 Ok(())
             }