+++ amethyst_renderer/src/resources.rs
@@ -29,6 +29,7 @@ impl<'a> PrefabData<'a> for AmbientColor {
         _: Entity,
         ambient: &mut Self::SystemData,
         _: &[Entity],
+        _: &[Entity],
     ) -> Result<(), Error> {
         ambient.0 = self.0;
         Ok(())